Initial commit
I have absolutely no interest in getting a SBT project to work nicely inside a
lein project, and I want to use ENSIME for this book, so I'm spinning it off
into its own project.

The tail-recursive implementation of the Fibonacci sequence eluded me because I
didn't realize that you have to make the current AND previous number a parameter
of the tail-position function.